fix: move sanitizer tests into test_api_telemetry.py to preserve xdist distribution
Merging the stack trace sanitizer tests into the existing telemetry test
file avoids adding a new test file, which would change pytest-xdist's
worker distribution and expose a pre-existing test ordering flake in
test_cli_bundle_explicit_parameters on Windows Python 3.9.
